FEAT : LineFactory 를 통한 생성자 내 비즈니스 로직 제거
